lab/analysis/requirements.txt
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
